REFACTOR: move wandb funcs from yolo_utils to wandb_utils
We want WandB to be a completely modular component. So we do not want it
to be named in any other script than inside the wandb_utils.py.

This commit moves 2 more functions from the yolo_utils to wandb_utils.
However, the few times wandb is named in the yolo_utils, and 1 time in
the MLProjectProcessor, are such specific usecases that I am not sure
yet if it should be moved. So they are left where they are, and we need
to look over them if this is functionality that we want to keep.
