+++ kso_utils/yolo_utils.py
@@ -927,49 +927,6 @@ def setup_paths(output_folder: str, model_type: str):
         return None, None
 
 
-def set_config(**kwargs):
-    """
-    `set_config` takes in a confidence threshold, model name, and evaluation directory and returns a
-    configuration object.
-
-    :param conf_thres: This is the confidence threshold for the bounding boxes
-    :type conf_thres: float
-    :param model: The name of the model you want to use
-    :type model: str
-    :param eval_dir: The directory where the evaluation images are stored
-    :type eval_dir: str
-    :return: The config object is being returned.
-    """
-    config = wandb.config
-    for key, value in kwargs.items():
-        if key == "model_name" and "model_name" in config:
-            pass
-        else:
-            setattr(config, key, value)
-    return config
-
-
-def add_data(path: str, name: str, registry: str, run):
-    """
-    > The function `add_data` takes a path to a directory, a name for the directory, and a run
-    object, and adds the directory to the run as an artifact
-
-    :param path: the path to the directory you want to upload
-    :type path: str
-    :param name: The name of the artifact
-    :type name: str
-    :param run: The run object that you get from calling wandb.init()
-    """
-    if registry == "wandb":
-        my_data = wandb.Artifact(name, type="raw_data")
-        if Path(path).is_dir():
-            my_data.add_dir(path)
-            run.log_artifact(my_data)
-        else:
-            my_data.add_file(path)
-            run.log_artifact(my_data)
-
-
 def generate_csv_report(
     evaluation_path: str,
     movie_csv_df: pd.DataFrame,
